# Improves AP startup time and CPU Usage