--- a/Code/Tools/AssetProcessor/native/resourcecompiler/RCCommon.cpp
+++ b/Code/Tools/AssetProcessor/native/resourcecompiler/RCCommon.cpp
@@ -59,6 +59,14 @@ namespace AssetProcessor
             );
     }
 
+    bool QueueElementID::operator!=(const QueueElementID& other) const
+    {
+        // if this becomes a hotspot in profile, we could use CRCs or other boost to comparison here.  These classes are constructed rarely
+        // compared to how commonly they are compared with each other.
+        return !(operator==(other));
+    }
+
+
     bool QueueElementID::operator<(const QueueElementID& other) const
     {
         int compare = m_sourceAssetReference.AbsolutePath().Compare(other.m_sourceAssetReference.AbsolutePath());