--- a/Code/Tools/AssetProcessor/native/resourcecompiler/RCCommon.h
+++ b/Code/Tools/AssetProcessor/native/resourcecompiler/RCCommon.h
@@ -29,6 +29,7 @@ namespace AssetProcessor
         void SetJobDescriptor(QString jobDescriptor);
         bool operator==(const QueueElementID& other) const;
         bool operator<(const QueueElementID& other) const;
+        bool operator!=(const QueueElementID& other) const;
 
     protected:
         SourceAssetReference m_sourceAssetReference;