--- a/Code/Tools/AssetProcessor/native/resourcecompiler/RCQueueSortModel.h
+++ b/Code/Tools/AssetProcessor/native/resourcecompiler/RCQueueSortModel.h
@@ -54,6 +54,9 @@ namespace AssetProcessor
         bool filterAcceptsRow(int source_row, const QModelIndex& source_parent) const override;
         bool lessThan(const QModelIndex& left, const QModelIndex& right) const override;
 
+        // Debugging functions
+        void DumpJobListInSortOrder();
+
     public Q_SLOTS:
         void OnEscalateJobs(AssetProcessor::JobIdEscalationList jobIdEscalationList);