--- a/Code/Tools/AssetProcessor/native/resourcecompiler/rccontroller.cpp
+++ b/Code/Tools/AssetProcessor/native/resourcecompiler/rccontroller.cpp
@@ -8,33 +8,84 @@
 
 #include "rccontroller.h"
 #include <native/resourcecompiler/RCCommon.h>
+#include <AzCore/std/parallel/thread.h>
 #include <QTimer>
 #include <QThreadPool>
 
-
-
 namespace AssetProcessor
 {
-    RCController::RCController(int cfg_minJobs, int cfg_maxJobs, QObject* parent)
+    void RCController::UpdateAndComputeJobSlots()
+    {
+        if (auto settingsRegistry = AZ::SettingsRegistry::Get())
+        {
+            auto settingsRoot = AZ::SettingsRegistryInterface::FixedValueString(AssetProcessorSettingsKey);
+            AZ::s64 valueFromRegistry = 0;
+            if (settingsRegistry->Get(valueFromRegistry, settingsRoot + "/Jobs/maxJobs"))
+            {
+                m_maxJobs = aznumeric_cast<int>(valueFromRegistry);
+            }
+
+            settingsRegistry->Get(m_alwaysUseMaxJobs, settingsRoot + "/Jobs/AlwaysUseMaxJobs");
+        }
+
+        if (m_maxJobs <= 1) // its not set in the registry to a specific value, or the registry set it to 0 (auto)
+        {
+            // Determine a good starting value for max jobs, we want to use hand tuned numbers for 2, 4, 8, 12, 16, etc
+            unsigned int cpuConcurrency = AZStd::thread::hardware_concurrency();
+            if (cpuConcurrency <= 1)
+            {
+                AZ_Printf(
+                    ConsoleChannel,
+                    "Unable to determine the number of hardware threads supported on this platform, assuming 4.\n",
+                    cpuConcurrency);
+                cpuConcurrency = 4; // we can't query it on this platform, set a reasonable default that gets some work done
+            }
+            AZ_Printf(ConsoleChannel, "Auto (0) selected for maxJobs - auto-configuring based on %u available CPU cores.\n", cpuConcurrency)
+
+            // for very low numbers of cores, hand-tune the values, these might be logical cores (hyperthread) and not real ones.
+            // we will reserve about half of this for "backround processing" and then the other half will be reserved for on-demand
+            // (critical or escalated) processing when we actually dispatch jobs.
+            if (cpuConcurrency <= 4)
+            {
+                m_maxJobs = 3;
+            }
+            else if (cpuConcurrency <= 6)
+            {
+                m_maxJobs = 5;
+            }
+            else
+            {
+                // for larger number of cores, 8, 16, 24, we want a few extra cores free
+                m_maxJobs = (cpuConcurrency - 2);
+            }
+        }
+
+        // final fail-safe
+        if (m_maxJobs < 2)
+        {
+            m_maxJobs = 2;
+        }
+        AZ_Printf(ConsoleChannel, "Asset Processor CPU Usage: (settings registry 'Jobs' section):\n")
+        AZ_Printf(ConsoleChannel, "    - Process up to %u jobs in parallel\n", m_maxJobs);
+        if (m_alwaysUseMaxJobs)
+        {
+            AZ_Printf(ConsoleChannel, "    - use all %u jobs whenever possible\n", m_maxJobs);
+        }
+        else
+        {
+            AZ_Printf(ConsoleChannel, "    - only use %u jobs when critical work is waiting, %u otherwise.\n", m_maxJobs, AZStd::GetMax(m_maxJobs / 2u, 1u));
+        }
+
+    }
+
+    RCController::RCController(QObject* parent)
         : QObject(parent)
         , m_dispatchingJobs(false)
         , m_shuttingDown(false)
     {
         AssetProcessorPlatformBus::Handler::BusConnect();
 
-        // Determine a good starting value for max jobs
-        int maxJobs = QThread::idealThreadCount();
-        if (maxJobs == -1)
-        {
-            maxJobs = 3;
-        }
-
-        maxJobs = qMax<int>(maxJobs - 1, 1);
-
-        // if the user has specified max jobs in the cfg file, then we obey their request
-        // regardless of whether they have chosen something bad or not - they would have had to explicitly
-        // pick this value (we ship with default 0 meaning auto), so if they've changed it, they intend it that way
-        m_maxJobs = cfg_maxJobs ? qMax(cfg_minJobs, cfg_maxJobs) :  maxJobs;
+        UpdateAndComputeJobSlots();
 
         m_RCQueueSortModel.AttachToModel(&m_RCJobListModel);
 
@@ -52,11 +103,6 @@ namespace AssetProcessor
     {
         AssetProcessorPlatformBus::Handler::BusDisconnect();
         m_RCQueueSortModel.AttachToModel(nullptr);
-    }
-
-    RCJobListModel* RCController::GetQueueModel()
-    {
-        return &m_RCJobListModel;
     }
 
     void RCController::StartJob(RCJob* rcJob)
@@ -334,22 +380,42 @@ namespace AssetProcessor
         if (!m_dispatchingJobs)
         {
             m_dispatchingJobs = true;
-            RCJob* rcJob = m_RCQueueSortModel.GetNextPendingJob();
 
-            while (m_RCJobListModel.jobsInFlight() < m_maxJobs && rcJob && !m_shuttingDown)
+            do
             {
-                if (m_dispatchingPaused)
+                RCJob* rcJob = m_RCQueueSortModel.GetNextPendingJob();
+
+                if (!rcJob)
                 {
-                    // note, even if dispatching is "paused" we start all "auto fail jobs" so that user gets instant feedback on failure.
-                    if (!rcJob->IsAutoFail())
+                    // there aren't any jobs remaining to dispatch.
+                    break;
+                }
+
+                // note that critical jobs and escalated jobs will always be at the top of the list
+                bool criticalOrEscalated = rcJob->IsCritical() || (rcJob->JobEscalation() > AssetProcessor::DefaultEscalation);
+
+                // do we have an open slot for this job?
+                unsigned int numJobsInFlight = m_RCJobListModel.jobsInFlight();
+                unsigned int regularJobLimit = m_alwaysUseMaxJobs ? m_maxJobs : AZStd::GetMax(m_maxJobs / 2, 1u);
+                unsigned int maxJobsToStart = criticalOrEscalated ? m_maxJobs : regularJobLimit;
+
+                // note that "auto fail jobs" oimmediately return as failed without doing any processing
+                // so they get to skip the line (they don't use up a thread
+                bool isAutoJob = rcJob->IsAutoFail();
+                bool tooManyJobs = numJobsInFlight >= maxJobsToStart;
+
+                if (!isAutoJob)
+                {
+                    if ((tooManyJobs) || (m_dispatchingPaused))
                     {
+                        // already using too much slots.
                         break;
                     }
                 }
                 StartJob(rcJob);
-                rcJob = m_RCQueueSortModel.GetNextPendingJob();
-            }
-            m_dispatchingJobs = false;
+            } while (true);
+            
+             m_dispatchingJobs = false;
         }
     }
     void RCController::DispatchJobs()
@@ -423,8 +489,9 @@ namespace AssetProcessor
             // PerformHeursticSearch already prints out the results.
             m_RCQueueSortModel.OnEscalateJobs(escalationList);
         }
-        // do not print a warning out when this fails, its fine for things to escalate jobs as a matter of course just to "make sure" they are escalated
-        // and its fine if none are in the build queue.
+
+        // escalating a job could free up an idle cpu thats dedicated to critical or escalated jobs.
+        DispatchJobs();
     }
 
     void RCController::OnEscalateJobsBySourceUUID(QString platform, AZ::Uuid sourceUuid)
@@ -445,6 +512,10 @@ namespace AssetProcessor
         }
         // do not print a warning out when this fails, its fine for things to escalate jobs as a matter of course just to "make sure" they are escalated
         // and its fine if none are in the build queue.
+
+        
+        // escalating a job could free up an idle cpu thats dedicated to critical or escalated jobs.
+        DispatchJobs();
     }
 
     void RCController::OnJobComplete(JobEntry completeEntry, AzToolsFramework::AssetSystem::JobStatus state)