--- a/Code/Tools/AssetProcessor/native/unittests/RCcontrollerUnitTests.cpp
+++ b/Code/Tools/AssetProcessor/native/unittests/RCcontrollerUnitTests.cpp
@@ -31,8 +31,6 @@ namespace
     constexpr int MaxProcessingWaitTimeMs = 60 * 1000; // Wait up to 1 minute.  Give a generous amount of time to allow for slow CPUs
     const ScanFolderInfo TestScanFolderInfo("c:/samplepath", "sampledisplayname", "samplekey", false, false);
     const AZ::Uuid BuilderUuid = AZ::Uuid::CreateRandom();
-    constexpr int MinRCJobs = 1;
-    constexpr int MaxRCJobs = 4;
 }
 
 class MockRCJob
@@ -233,7 +231,7 @@ void RCcontrollerUnitTests::SetUp()
 {
     UnitTest::AssetProcessorUnitTestBase::SetUp();
 
-    m_rcController = AZStd::make_unique<AssetProcessor::RCController>(MinRCJobs, MaxRCJobs);
+    m_rcController = AZStd::make_unique<AssetProcessor::RCController>();
 
     QDir assetRootPath(m_assetDatabaseRequestsHandler->GetAssetRootDir().c_str());