… Code/Tools/AssetProcessor/native/tests/assetmanager/AssetManagerTestingBase.cpp
… Code/Tools/AssetProcessor/native/tests/assetmanager/AssetManagerTestingBase.h
… Code/Tools/AssetProcessor/native/tests/assetmanager/AssetProcessorManagerTest.cpp
… Code/Tools/AssetProcessor/native/tests/resourcecompiler/RCControllerTest.cpp
… Code/Tools/AssetProcessor/native/ui/MainWindow.ui
… Code/Tools/AssetProcessor/native/utilities/ApplicationManager.cpp
… Code/Tools/AssetProcessor/native/utilities/GUIApplicationManager.cpp
… Code/Tools/AssetProcessor/native/utilities/PlatformConfiguration.cpp
… Code/Tools/AssetProcessor/native/utilities/PlatformConfiguration.h
… Code/Tools/AssetProcessor/native/utilities/assetUtils.cpp
… Code/Tools/AssetProcessor/native/utilities/assetUtils.h
… Gems/Atom/RPI/Code/Include/Atom/RPI.Reflect/Asset/AssetUtils.h
… Gems/Atom/RPI/Code/Source/RPI.Builders/Material/MaterialBuilder.cpp
… Gems/Atom/RPI/Code/Source/RPI.Builders/Material/MaterialBuilderUtils.cpp
… Gems/Atom/RPI/Code/Source/RPI.Builders/Material/MaterialBuilderUtils.h
… Gems/Atom/RPI/Code/Source/RPI.Builders/Material/MaterialTypeBuilder.cpp
… Gems/Atom/RPI/Code/Source/RPI.Reflect/Asset/AssetUtils.cpp
… Gems/Atom/RPI/Code/Source/RPI.Reflect/Model/ModelAsset.cpp
… Registry/settings.assetprocessorbatch.setreg