## What does this PR do?
Its a basic rework of how the user options in AssetProcessor work (move away from QSettings, which is a global settings options that goes in your windows registry, to SetReg instead).

### User Settings changes:
* Asset Processor `AssetUtilities::` now has a Set User Setting, Get User Setting, and a change notify for user settings, utility functions.
* Existing settings on the `settings` page moved over to this new system to avoid bleeding settings across projects
* new "Log verbosity" option to stop logs from becoming hundreds of megabytes long unless you want it.  Takes effect instantly.
* Minjobs (useless setting) removed.
* MaxJobs setting retained but its now read by the thing that uses it instead (the actual job queue controller that starts jobs).
![image](https://github.com/user-attachments/assets/6be20aa3-1d4d-41e2-8f54-420fc04d37a0)

### Job Queue changes
* MaterialBuilder no longer specifies that all textures are job dependencies of materials (this was unnecessary and made it so that textures were escalated above other asset types).
* AssetProcessor escalates jobs that are blocking critical jobs from starting instead of just waiting for them to happen naturally.
* AssetProcessor reserves half jof the CPU cores for critical jobs.  When no critical jobs are running, it uses fewer cores to give more CPU usage to the editor to maintain framerates.  Batch version keeps using all cpus.
* Setting added to make AP use all cores instead, and that setting is set to true in the AssetProcessorBatch, false (default) in all other situations.
* Escalating a job (so when the editor asks for you to compile an asset immediately) causes a dispatch of the job queue, so that it can immediately use up those reserved slots.
* Common platform jobs are now critical jobs by default, because they are jobs which generate other jobs, which themselves could be critical.

### Misc changes
* Utility function to dump the job queue for debugging
* Cleaning up incorrect comments or outputs
* Removing using Qt to discover cores in favor of std.
* Renamed an enum value in AssetProcessor namespace from `AssetProcessor::Default` to `AssetProcessor::DefaultEscalation` - `Default` is not a good name for something in that entire namespace.  (not an enum class!)
* Sped up escalation by adding a escalate-by-queueid function
* Sped up escalation by not escalating jobs already escalated.

Timing WITH new changes (starting editor with no cache)
  *   42s (  42s total) analyzing
  * 2m45s (3m28s total) "Asset Processor working..."
  * 1m15s (4m44s total) 60fps interactive editor with defaultlevel (meshes still compiling)

Timing WITHOUT the new changes (starting editor with no cache)
  *   42s (  42s total) analyzing (I did not change analyze so this makes sense)
  * 3m34s (4m18s total) "Asset processor working..."
  * 1m22s (5m40s total) 10-20 fps interactive editor with defaultlevel (meshes still compiling)

So about a 45s saving on "Asset Processor Working" and about 56s total saving with a 60fps editor while it works in background.

Both timings were captured with all other variables hte same, and were performed on a "hot" file cache, as in, run it once, then clear cache
and immediately run it again to get official timings.    The timing
on a cold cache for the new changes is even more of a difference than
a cold cache without the changes.

## How was this PR tested?

Starting AP from scratch repeatedly and opening up the default level as soon as its possible to do so, making sure no crash, and that the level eventually fully loads.