@nick-l-o3de 2025-06-14
I've implemented the change suggestions, and also found the tests were broken by them.

New commit includes all PR suggestions as well as fix for tests.

Running AssetProcessor.Tests
```
[----------] Global test environment tear-down
[==========] 680 tests from 79 test suites ran. (147262 ms total)
[  PASSED  ] 680 tests.
```